backend/requirements.txt
flask
flask-cors
streamlit
requests
groq
PyJWT
werkzeug
